+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 - `InteractionGroups` constructor now requires an `InteractionTestMode` parameter. If you want same behaviour as before, use `InteractionTestMode::And` (eg. `InteractionGroups::new(Group::GROUP_1, Group::GROUP_1, InteractionTestMode::And)`)
 - `CoefficientCombineRule::Min` - now makes sure it uses a non zero value as result by using `coeff1.min(coeff2).abs()`
 - `InteractionTestMode`: Specifies which method should be used to test interactions. Supports `AND` and `OR`.
-- `CoefficientCombineRule::Sum` - Adds the two coefficients and does a clamp to have at most 1.
+- `CoefficientCombineRule::ClampedSum` - Adds the two coefficients and does a clamp to have at most 1.
 
 ## v0.30.1 (17 Oct. 2025)